requirements.txt
praw
langchain
crewai
telebot
python-dotenv
pyTelegramBotAPI
markdown
telegraph
BeautifulSoup
streamlit
streamlit-quill
flask